consolidate execution context files in the execution_algorithms directories
