+++ include/lbann/CMakeLists.txt
@@ -12,7 +12,6 @@ add_subdirectory(callbacks)
 add_subdirectory(data_coordinator)
 add_subdirectory(data_readers)
 add_subdirectory(data_store)
-add_subdirectory(execution_contexts)
 add_subdirectory(io)
 add_subdirectory(layers)
 add_subdirectory(macros)
