--- include/lbann/execution_contexts/CMakeLists.txt
@@ -1,8 +1,0 @@
-# Add the headers for this directory
-set_full_path(THIS_DIR_HEADERS
-  execution_context.hpp
-  sgd_execution_context.hpp
-  )
-
-# Propagate the files up the tree
-set(HEADERS "${HEADERS}" "${THIS_DIR_HEADERS}" PARENT_SCOPE)
