+++ src/CMakeLists.txt
@@ -10,7 +10,6 @@ add_subdirectory(callbacks)
 add_subdirectory(data_coordinator)
 add_subdirectory(data_readers)
 add_subdirectory(data_store)
-add_subdirectory(execution_contexts)
 add_subdirectory(io)
 add_subdirectory(layers)
 add_subdirectory(metrics)
