+++ src/layers/data_type_layer.cpp
@@ -29,7 +29,7 @@
 
 #include "matrix_builder.hpp"
 
-#include "lbann/execution_contexts/sgd_execution_context.hpp"
+#include "lbann/execution_algorithms/sgd_execution_context.hpp"
 #include "lbann/layers/data_type_layer.hpp"
 #include "lbann/models/model.hpp"
 #include "lbann/trainers/trainer.hpp"
